requirements.txt
h5py>=3.8
numpy>=1.24
PyYAML>=6.0
tensorboard>=2.14
torch>=2.1
torchvision>=0.16
tqdm>=4.66
robomimic>=0.3
robosuite>=1.4
